2016-2017 Social Details, Better About Pages, and Award Winners
The new social page has better text descriptions of events from
2016-2017 with supporting pictures. A typo was fixed in the Racing
bylaws. The main page was updated with a more accurate description of
the current student branch and a few committees. The upcoming events
were clarified further. The about page received a long-needed major
overhaul that made it more comprehensive of the organization today
instead of in Spencer’s day. Finally, the award winners from last year
were properly recognized as they should be in order to meet IEEE
criteria for specific recognition.