Terms for Racing Committee Chairs and Consistency
Erick was placed before Adrian and listed as the Chair for Summer 2017,
and Adrian was listed as the Chair for Fall 2017 and Spring 2018.
Computer Society is added as a past involvement for Zach. Small
stylistic changes were made to the biographies to make them consistent
with each other with updates to modern IEEE terminology.